feat(types): move event stream types to eventStream file
