+++ packages/types/src/eventStream.ts
@@ -1,3 +1,5 @@
+import { RequestSigner } from "./signature";
+import { HttpRequest } from "./http";
 /**
  * An event stream message. The headers and body properties will always be
  * defined, with empty headers represented as an object with no keys and an
@@ -72,4 +74,31 @@ export interface Int64 {
   readonly bytes: Uint8Array;
   valueOf: () => number;
   toString: () => string;
+}
+
+/**
+ * Util functions for serializing or deserializing event stream
+ */
+export interface EventStreamSerdeContext {
+  eventStreamMarshaller: EventStreamMarshaller;
+}
+
+export interface EventStreamMarshaller {
+  deserialize: (
+    body: any,
+    deserializer: (msg: Message) => any,
+    exceptionsDeserializer: (msg: Message) => any
+  ) => AsyncIterable<any>;
+  serialize: (
+    input: AsyncIterable<any>,
+    serializer: (event: any) => Message
+  ) => any;
+}
+
+export interface EventStreamRequestSigner {
+  sign(request: HttpRequest): Promise<HttpRequest>;
+}
+
+export interface EventStreamSerdeProvider {
+  (options: any): [EventStreamMarshaller, EventStreamRequestSigner];
 }
